LayoutTests/fast/workers/worker-cloneport-expected.txt

CONSOLE MESSAGE: line 16: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test MessagePort messaging/entangle/detangle across threads. Should print "SUCCESS" when done.
